# backend/advanced/documentation/swagger.yaml
openapi: 3.0.0
info:
  title: Todo List API
  version: 1.0.0
  description: API for managing todo items

servers:
  - url: http://localhost:8000

paths:
  /todos:
    get:
      summary: Get all todos
      description: Retrieve a list of all todo items
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      summary: Create a new todo
      description: Create a new todo item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
              required:
                - description
      responses:
        "201":
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /todos/{id}:
    get:
      summary: Get a specific todo
      description: Retrieve a specific todo item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    put:
      summary: Update a todo
      description: Update a specific todo item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                completed:
                  type: boolean
      responses:
        "200":
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    delete:
      summary: Delete a todo
      description: Delete a specific todo item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Todo deleted successfully
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - description
        - completed
      properties:
        id:
          type: integer
          description: The unique ID of the todo.
        description:
          type: string
          description: The description of the todo.
        completed:
          type: boolean
          description: Indicates if the todo is completed or not.